Remove redundant StateManager class
