+++ src/sneck/classes/game.py
@@ -1,10 +1,12 @@
 from sneck.classes.board import Board
+from sneck.classes.game_states import (GameOverState, PlayingState,
+                                       ScoreBoardState, TitleScreenState)
 from sneck.classes.score_board_data import ScoreBoardData
 from sneck.classes.screen import Screen
-from sneck.classes.state_manager import StateManager
+from sneck.protocols.game_context import GameContext
 
 
-class Game:
+class Game(GameContext):
     FPS = 7
     ROWS = 15
     COLS = 22
@@ -16,8 +18,21 @@ class Game:
 
         self.board = Board(self.ROWS, self.COLS)
         self.screen = Screen(self.ROWS, self.COLS)
-        self.state_manager = StateManager(self)
+
+        self.state = TitleScreenState(self)
 
     def run(self) -> None:
         while True:
-            self.state_manager.run()
+            self.state.run()
+
+    def transition_to_title_screen(self):
+        self.state = TitleScreenState(self)
+
+    def transition_to_playing(self):
+        self.state = PlayingState(self)
+
+    def transition_to_game_over(self):
+        self.state = GameOverState(self)
+
+    def transition_to_score_board(self):
+        self.state = ScoreBoardState(self)
